6-serverless-xray-stack/requirements.txt
aws_cdk.core==1.27.0
aws_cdk.aws_ec2==1.27.0
aws_cdk.aws_ecs==1.27.0
aws_cdk.aws_cloudformation==1.27.0
aws_cdk.aws_lambda==1.27.0
aws_cdk.aws_s3==1.27.0
aws_cdk.aws_s3_deployment==1.27.0
aws_cdk.aws_dynamodb==1.27.0
aws_cdk.aws_elasticloadbalancingv2==1.27.0
